+++ extras/api/auto_llama_api/__init__.py
@@ -1,0 +1,2 @@
+from .__settings__ import settings, auto_llama_config
+from .app import main, app
