--- extras/api/src/__init__.py
@@ -1,2 +1,0 @@
-from .__settings__ import settings
-from .app import main, app
