--- extras/api/src/__settings__.py
@@ -1,13 +1,0 @@
-from pathlib import Path
-
-from pydantic_settings import BaseSettings, SettingsConfigDict
-
-
-class Settings(BaseSettings):
-    model_config = SettingsConfigDict(
-        env_file=Path(__file__).parent.parent / ".env",
-        case_sensitive=True,
-    )
-
-
-settings = Settings.model_validate({})
